Rename scripts to ovena-*

+++ bin/ovena-db_restore.sh
@@ -3,6 +3,7 @@
 # v0.2.0
 #
 # Restore orthanc PostgreSQL DB.
+# If no backup file is passed, will create the database user and empty database, used for initial DB setup.
 #
 
 BKPFILE="$1"
@@ -11,10 +12,14 @@ DBUSER="$<DATABASE_USERNAME>"
 alias _exec="docker-compose -f $<OVENA_CONFIG>/docker-compose.yml exec \"$<DATABASE_DOCKER_IMAGE>\""
 USEROPT="--username=postgres"
 
-_exec /usr/bin/dropdb $USEROPT "$DBNAME"
-_exec /usr/bin/dropuser $USEROPT "$DBUSER"
+if [ -f "$BKPFILE" ]; then
+    _exec /usr/bin/dropdb $USEROPT "$DBNAME"
+    _exec /usr/bin/dropuser $USEROPT "$DBUSER"
+fi
 
 _exec /usr/bin/createuser $USEROPT "$DBUSER" || exit
 _exec /usr/bin/createdb $USEROPT --owner="$DBUSER" "$DBNAME" || exit
 
-zcat "$1" | _exec /usr/bin/psql --username="$DBUSER" 
+if [ -f "$BKPFILE" ]; then
+    zcat "$1" | _exec /usr/bin/psql --username="$DBUSER"
+fi
